feat: add jest-watch package
