+++ packages/jest-cli/src/__tests__/watch_test_name_pattern_mode.test.js
@@ -9,7 +9,7 @@
 'use strict';
 
 import chalk from 'chalk';
-import {KEYS} from '../constants';
+import {KEYS} from 'jest-watch';
 
 const runJestMock = jest.fn();
 
@@ -92,10 +92,6 @@ jest.doMock(
       return Promise.resolve();
     },
 );
-
-jest.doMock('../lib/terminal_utils', () => ({
-  getTerminalWidth: () => terminalWidth,
-}));
 
 const watch = require('../watch').default;
