+++ packages/jest-cli/src/plugins/update_snapshots_interactive.js
@@ -6,10 +6,10 @@
  *
  * @flow
  */
-import type {JestHookSubscriber} from '../jest_hooks';
+import type {JestHookSubscriber} from 'types/JestHooks';
 import type {GlobalConfig} from 'types/Config';
 import type {AggregatedResult, AssertionLocation} from 'types/TestResult';
-import BaseWatchPlugin from '../base_watch_plugin';
+import {BaseWatchPlugin} from 'jest-watch';
 import SnapshotInteractiveMode from '../snapshot_interactive_mode';
 
 class UpdateSnapshotInteractivePlugin extends BaseWatchPlugin {
