+++ packages/jest-watch/src/index.js
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+export {default as BaseWatchPlugin} from './base_watch_plugin';
+export {default as JestHook} from './jest_hooks';
+export {default as PatternPrompt} from './pattern_prompt';
+export * from './constants';
+export {default as Prompt} from './lib/Prompt';
+export * from './lib/pattern_mode_helpers';
