+++ packages/jest-watch/src/lib/__tests__/prompt.test.js
