+++ packages/jest-watch/src/lib/colorize.js
