+++ packages/jest-watch/src/lib/format_test_name_by_pattern.js
