+++ packages/jest-watch/src/lib/pattern_mode_helpers.js
